Move all basic stuff to craftassist agent test dir
