+++ agents/craftassist/tests/base_craftassist_test_case.py
