+++ agents/craftassist/tests/fake_agent.py
