+++ agents/craftassist/tests/fake_mobs.py
