+++ agents/craftassist/tests/interactive_memory_explorer.py
