+++ agents/craftassist/tests/mc_agent.py
