+++ agents/craftassist/tests/test.sh
