+++ agents/craftassist/tests/test_agent.py
