+++ agents/craftassist/tests/visualize_scenario.py
