+++ agents/craftassist/tests/world.py
