+++ agents/craftassist/tests/world_visualizer.py
