+++ droidlet/dialog/craftassist/tests/test_dialogue_manager.py
@@ -9,7 +9,7 @@ from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.dialog.droidlet_nsp_model_wrapper import DroidletNSPModelWrapper
 from agents.loco_mc_agent import LocoMCAgent
 from droidlet.interpreter.tests.all_test_commands import *
-from craftassist.test.fake_agent import MockOpt
+from agents.craftassist.tests.fake_agent import MockOpt
 
 
 class AttributeDict(dict):
