+++ droidlet/dialog/craftassist/tests/test_safety.py
@@ -5,8 +5,8 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import os
 import unittest
 
-from craftassist.test.base_craftassist_test_case import BaseCraftassistTestCase
-from craftassist.test.fake_agent import MockOpt
+from agents.craftassist.tests.base_craftassist_test_case import BaseCraftassistTestCase
+from agents.craftassist.tests.fake_agent import MockOpt
 
 TTAD_BERT_DATA_DIR = os.path.join(os.path.dirname(__file__), "../../../../craftassist/agent/datasets/annotated_data/")
 TTAD_BERT_MODEL_DIR = os.path.join(os.path.dirname(__file__), "../../../../craftassist/agent/models/semantic_parser/")
