+++ droidlet/dialog/craftassist/tests/test_y_print_parsing_report.py
@@ -8,7 +8,7 @@ import json
 from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.dialog.droidlet_nsp_model_wrapper import DroidletNSPModelWrapper
 from agents.loco_mc_agent import LocoMCAgent
-from craftassist.test.fake_agent import MockOpt
+from agents.craftassist.tests.fake_agent import MockOpt
 from prettytable import PrettyTable
